Refactor physics PMX test fixtures
